refactor: calculateCarItems 함수 모듈화 및 컨벤션 반영
